snippets/apps/user_snippets/serializers/tags.py
from rest_framework import serializers

from apps.user_snippets.models import Tag


class TagSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tag
        fields = ('id', 'title')



